+++ src/module/ColorCode.js
@@ -1,3 +1,10 @@
+/**
+ * History
+ ** 2023-03-25: 임주형 - 최초생성
+ * 설명
+ ** 색상 코드를 반환하는 함수 제공
+ */
+
 const MemberGrade = {
     LV1: "LV1",
     LV2: "LV2",
